refactor: move MockUI to root package (CP: 25.0)
